Rename /first-time-login to /join and FirstTimeLoginPage to JoinPage
