+++ assets/js/pages/JoinPage/index.tsx
